// src/interfaces/IPessoas.ts
export interface IPessoas {
  nome_completo: string
  email: string;
  telefone_celular: string
  placa: string;
  marca: string;
  cor: string;
  modelo: string;
  cpf: string;
  rg: string;
  matricula: string;
  data_contratacao: string;
  data_demissao: string;
  data_inicial_ferias: string;
  data_final_ferias: string;
  observacoes: string;
  departamento_id: number;
  departamento_nome: string;
  departamento_numero: number;
  local_especifico: string;
  tag_hexa: string;
  tag_dec: string
}